sliders mostly working
